add server side rendered pages protection and api route protection

+++ pages/api/products/[id].js
@@ -1,6 +1,8 @@
 import dbConnect from "../../../lib/dbConnect";
 import Product from "../../../models/Product";
 import { getProductById } from "../../../services/productService";
+import { unstable_getServerSession } from "next-auth/next";
+import { authOptions } from "../auth/[...nextauth]";
 
 export default async function handler(request, response) {
   const { id } = request.query;
@@ -11,12 +13,22 @@ export default async function handler(request, response) {
       const product = await getProductById(id);
       return response.status(200).json(product);
     case "PUT":
+      if (!(await unstable_getServerSession(request, response, authOptions))) {
+        return response.status(401).json({
+          error: "You must be signed in to update products.",
+        });
+      }
       const data = JSON.parse(request.body);
       const updatedProduct = await Product.findByIdAndUpdate(id, data);
       return response
         .status(200)
         .json({ message: "Product updated", updatedProduct: updatedProduct });
     case "DELETE":
+      if (!(await unstable_getServerSession(request, response, authOptions))) {
+        return response.status(401).json({
+          error: "You must be signed in to delete products.",
+        });
+      }
       await Product.findByIdAndDelete(id);
       return response
         .status(200)
